Avoid stripping of trailing space for 'fillchars'
The 'fillchars' setting in this ftplugin is a space character. Therefore
it ends with a space character. This makes it a bit fragile when
trimming all trailing whitespace charaters.
Therefore a comment was appended to avoid that the space character is
the last character in the line and will not get stripped automatically.